pyproject.toml
[project]
name = "rag-production"
version = "0.1.0"
description = "Production RAG App with Hybrid Retrieval and Cross-Encoder Reranking"
authors = [
    {name = "User"}
]
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-groq",
    "langchain-huggingface",
    "sentence-transformers",
    "langchain-cohere",
    "pypdf",
    "pymupdf",
    "python-docx",
    "pandas",
    "rank_bm25",
    "langchain-chroma",
    "python-dotenv",
    "pytest",
    "streamlit"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
